requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.12
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.2
django-summernote==0.8.11.6
exceptiongroup==1.1.1
Faker==12.0.1
iniconfig==2.0.0
mixer==7.2.2
packaging==23.1
pluggy==1.0.0
pycodestyle==2.8.0
pytest==7.3.1
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
